Use Jest snapshots instead of files
