--- test/constructor-and-class-member-annotation.js
@@ -1,20 +1,0 @@
-/* @flow */
-import React from 'react';
-
-export default class Component extends React.Component {
-  constructor(props) {
-    super(props);
-  }
-
-  componentDidMount() {
-  }
-}
-
-class Component2 extends React.Component {
-  constructor(props) {
-    super(props);
-  }
-
-  componentDidMount() {
-  }
-}
