--- test/constructor-and-class-member-annotation.output.js
@@ -1,24 +1,0 @@
-/* @flow */
-import React from 'react';
-
-export default class Component extends React.Component {
-  constructor(props: Props) {
-    super(props);
-  }
-
-  props: Props;
-
-  componentDidMount() {
-  }
-}
-
-class Component2 extends React.Component {
-  constructor(props: Props) {
-    super(props);
-  }
-
-  props: Props;
-
-  componentDidMount() {
-  }
-}
