--- test/optional-proptypes.js
@@ -1,33 +1,0 @@
-import React, { PropTypes } from 'react';
-
-function Button(props) {
-  return (
-    <button>
-      {React.Children.toArray(children)}
-    </button>
-  );
-}
-
-Button.propTypes = {
-  optionalArray: PropTypes.array,
-  optionalBool: PropTypes.bool,
-  optionalFunc: PropTypes.func,
-  optionalNumber: PropTypes.number,
-  optionalObject: PropTypes.object,
-  optionalString: PropTypes.string,
-  optionalNode: PropTypes.node,
-  optionalElement: PropTypes.element,
-  optionalMessage: PropTypes.instanceOf(Message),
-  optionalEnum: PropTypes.oneOf(['News', 'Photos']),
-  optionalUnion: PropTypes.oneOfType([
-    PropTypes.string,
-    PropTypes.number,
-    PropTypes.instanceOf(Message)
-  ]),
-  optionalArrayOf: PropTypes.arrayOf(PropTypes.number),
-  optionalObjectOf: PropTypes.objectOf(PropTypes.number),
-  optionalObjectWithShape: PropTypes.shape({
-    color: PropTypes.string,
-    fontSize: PropTypes.number
-  }),
-};
