--- test/react-optional-proptypes.output.js
@@ -1,25 +1,0 @@
-/* @flow */
-import React from 'react';
-
-export type Props = {
-  optionalArray?: Array<any>,
-  optionalBool?: boolean,
-  optionalFunc?: Function,
-  optionalNumber?: number,
-  optionalObject?: Object,
-  optionalString?: string,
-  optionalNode?: number | string | React.Element | Array<any>,
-  optionalElement?: React.Element,
-  optionalMessage?: Message,
-  optionalEnum?: 'News' | 'Photos',
-  optionalUnion?: string | number | Message,
-  optionalArrayOf?: Array<number>,
-  optionalObjectOf?: Object<number>,
-  optionalObjectWithShape?: {
-    color?: string,
-    fontSize?: number,
-  },
-};
-
-export const F = (props: Props) =>
-	<div></div>;
